refactor: took ownership of the Controller class by renaming it (and related code) to OdrlController
