+++ controller/src/classes/accessRequests/OdrlAccessRequest.ts
@@ -4,7 +4,7 @@ import { Resources } from "@/types";
 import { getDefaultSession } from "@inrupt/solid-client-authn-browser";
 import { deleteFile } from "@inrupt/solid-client";
 
-export class InruptAccessRequest extends AccessRequest implements IAccessRequest {
+export class ODRLAccessRequest extends AccessRequest implements IAccessRequest {
     constructor(controller: IController<{}>, inboxConstructor: IInboxConstructor, resourcesStore: IStore<Resources>) {
         super(controller, inboxConstructor, resourcesStore)
     }
