+++ controller/src/classes/index.ts
@@ -1,5 +1,5 @@
-export * from './permissionManager/inrupt';
+export * from './permissionManager/odrl';
 export * from './stores/InruptStore';
 export * from './stores/BaseStore';
 export * from './subjectResolvers';
-export * from './Controller';
+export * from './OdrlController';
