+++ controller/src/controllers.ts
@@ -1,6 +1,6 @@
-import { Controller } from "./classes/Controller";
-import { PublicManager } from "./classes/permissionManager/inrupt/PublicManager";
-import { WebIdManager } from "./classes/permissionManager/inrupt/WebIdManager";
+import { ODRLController } from "./classes/OdrlController";
+import { PublicManager } from "./classes/permissionManager/odrl/PublicManager";
+import { WebIdManager } from "./classes/permissionManager/odrl/WebIdManager";
 import { InruptInboxStore } from "./classes/stores/InruptInboxStore";
 import { InruptStore } from "./classes/stores/InruptStore";
 import { PublicResolver } from "./classes/subjectResolvers/Public";
@@ -8,7 +8,7 @@ import { WebIdResolver } from "./classes/subjectResolvers/WebId";
 import { PublicSubject, WebIdSubject } from "./types/subjects";
 
 export const createBasicController = () => {
-    return new Controller<{
+    return new ODRLController<{
         webId: WebIdSubject,
         public: PublicSubject,
     }>(
